[-] support for ANSI encoded dictionary: fixed incorrect split by line

+++ branches/wx-port/src/Phrase.h
@@ -2,7 +2,7 @@
 
 struct Phrase
 {
-	long Priority;
+	int Priority;
 	wxString MatchExpr;
 	wxString Answer;
 	wxString Flags;
